+++ asynch/pool.py
@@ -2,10 +2,10 @@ import asyncio
 import logging
 from collections import deque
 from collections.abc import AsyncIterator
-from contextlib import asynccontextmanager
+from contextlib import asynccontextmanager, suppress
 from typing import Optional
 
-from asynch.connection import Connection, connect
+from asynch.connection import Connection
 from asynch.errors import AsynchPoolError
 from asynch.proto import constants
 from asynch.proto.models.enums import PoolStatus
@@ -33,8 +33,8 @@ class Pool:
         self._lock = asyncio.Lock()
         self._acquired_connections: deque[Connection] = deque(maxlen=maxsize)
         self._free_connections: deque[Connection] = deque(maxlen=maxsize)
-        self._opened: Optional[bool] = None
-        self._closed: Optional[bool] = None
+        self._opened: bool = False
+        self._closed: bool = False
 
     async def __aenter__(self) -> "Pool":
         await self.startup()
@@ -44,7 +44,7 @@ class Pool:
         await self.shutdown()
 
     def __repr__(self) -> str:
-        cls_name = self.__class__.__name__
+        cls_name = type(self).__name__
         status = self.status
         return (
             f"<{cls_name}(minsize={self._minsize}, maxsize={self._maxsize})"
@@ -52,65 +52,49 @@ class Pool:
         )
 
     @property
+    def opened(self) -> bool:
+        """Returns True if the pool is opened.
+
+        :returns: the pool open status
+        :rtype: bool
+        """
+
+        return self._opened
+
+    @property
+    def closed(self) -> bool:
+        """Return True if the pool is closed.
+
+        :returns: the pool close status
+        :rtype: bool
+        """
+
+        return self._closed
+
+    @property
     def status(self) -> str:
         """Return the status of the pool.
-
-        If pool.opened is None and pool.closed is None,
-        then the pool is in the "created" state.
-        It was neither opened nor closed.
-
-        When executing `async with pool: ...`,
-        the `pool.opened` is True and `pool.closed` is None.
-        When leaving the context, the `pool.closed` is True
-        and the `pool.opened` is False.
 
         :raise AsynchPoolError: an unresolved pool state.
         :return: the Pool object status
         :rtype: str (PoolStatus StrEnum)
         """
 
-        if self._opened is None and self._closed is None:
+        if not (self._opened or self._closed):
             return PoolStatus.created
-        if self._opened:
+        if self._opened and not self._closed:
             return PoolStatus.opened
-        if self._closed:
+        if self._closed and not self._opened:
             return PoolStatus.closed
         raise AsynchPoolError(f"{self} is in an unknown state")
 
     @property
-    def closed(self) -> Optional[bool]:
-        """Returns the pool close status.
-
-        If the return value is None,
-        the pool was only created,
-        but neither activated or closed.
-
-        :returns: the connection close status
-        :rtype: None | bool
-        """
-
-        return self._closed
-
-    @property
-    def connections(self) -> int:
-        """Returns the number of connections in the pool.
-
-        This number represents the current size of the pool
-        which is the sum of the acquired and free connections.
-
-        :return: the number of connections in the pool
-        :rtype: int
-        """
-
-        return self.acquired_connections + self.free_connections
-
-    @property
     def acquired_connections(self) -> int:
-        """Returns the number of connections acquired from the pool.
+        """Return the number of connections acquired from the pool.
 
-        A connection is acquired when `pool.connection()` is invoked.
+        A connection is acquired when the `pool.connection()` is invoked.
 
-        :return: the number of connections requested the pool
+        :return: the number of connections requested from the pool
         :rtype: int
         """
 
@@ -118,13 +102,27 @@ class Pool:
 
     @property
     def free_connections(self) -> int:
-        """Returns the number of free connections in the pool.
+        """Return the number of free connections in the pool.
 
-        :return: the number of connections in the pool
+        :return: the number of free connections in the pool
         :rtype: int
         """
 
         return len(self._free_connections)
+
+    @property
+    def _pool_size(self) -> int:
+        """Return the number of connections associated with the pool.
+
+        This number is the sum of the acquired and free connections.
+        So this sum may be interpreted as the current size of the pool
+        or the number of connections associated with the pool and so on.
+
+        :return: the number of connections related to the pool
+        :rtype: int
+        """
+
+        return self.acquired_connections + self.free_connections
 
     @property
     def maxsize(self) -> int:
@@ -135,20 +133,41 @@ class Pool:
         return self._minsize
 
     async def _create_connection(self) -> None:
-        pool_size, maxsize = self.connections, self.maxsize
-        if pool_size == maxsize:
+        if self._pool_size == self._maxsize:
             raise AsynchPoolError(f"{self} is already full")
-        if pool_size > maxsize:
+        if self._pool_size > self._maxsize:
             raise AsynchPoolError(f"{self} is overburden")
 
-        conn = await connect(**self._connection_kwargs)
-        self._free_connections.append(conn)
+        conn = Connection(**self._connection_kwargs)
+        await conn.connect()
 
-    async def _acquire_connection(self) -> Connection:
+        try:
+            await conn.ping()
+            self._free_connections.append(conn)
+        except ConnectionError as e:
+            msg = f"failed to create a {conn} for {self}"
+            raise AsynchPoolError(msg) from e
+
+    def _pop_connection(self) -> Connection:
         if not self._free_connections:
             raise AsynchPoolError(f"no free connection in {self}")
+        return self._free_connections.popleft()
 
-        conn = self._free_connections.popleft()
+    async def _get_fresh_connection(self) -> Optional[Connection]:
+        while self._free_connections:
+            conn = self._pop_connection()
+            with suppress(ConnectionError):
+                await conn._refresh()
+                return conn
+        return None
+
+    async def _acquire_connection(self) -> Connection:
+        if conn := await self._get_fresh_connection():
+            self._acquired_connections.append(conn)
+            return conn
+
+        await self._create_connection()
+        conn = self._pop_connection()
         self._acquired_connections.append(conn)
         return conn
 
@@ -157,22 +176,43 @@ class Pool:
             raise AsynchPoolError(f"the connection {conn} does not belong to {self}")
 
         self._acquired_connections.remove(conn)
+        try:
+            await conn._refresh()
+        except ConnectionError as e:
+            msg = f"the {conn} is invalidated"
+            raise AsynchPoolError(msg) from e
+
         self._free_connections.append(conn)
 
-    async def _init_connections(self, n: Optional[int] = None) -> None:
-        to_create = n if n is not None else self.minsize
-        if to_create < 0:
-            msg = f"cannot create ({to_create}) negative connections for {self}"
+    async def _init_connections(self, n: int, *, strict: bool = False) -> None:
+        if n < 0:
+            msg = f"cannot create a negative number ({n}) of connections for {self}"
             raise ValueError(msg)
-        if to_create == 0:
-            return
-        if (self.connections + to_create) > self.maxsize:
-            msg = f"cannot create {to_create} connections to exceed the size of {self}"
+        if (self._pool_size + n) > self.maxsize:
+            msg = (
+                f"{self} has the {self._pool_size} connections, "
+                f"adding {n} will exceed its maxsize ({self.maxsize})"
+            )
             raise AsynchPoolError(msg)
+        if not n:
+            return
+
+        # it is possible that the `_create_connection` may not create `n` connections
         tasks: list[asyncio.Task] = [
-            asyncio.create_task(self._create_connection()) for _ in range(to_create)
+            asyncio.create_task(self._create_connection()) for _ in range(n)
         ]
-        await asyncio.wait(fs=tasks)
+        # that is why possible exceptions from the `_create_connection` are also gathered
+        if strict and any(
+            i
+            for i in await asyncio.gather(*tasks, return_exceptions=True)
+            if isinstance(i, Exception)
+        ):
+            msg = f"failed to create the {n} connection(s) for the {self}"
+            raise AsynchPoolError(msg)
+
+    async def _ensure_minsize_connections(self, *, strict: bool = False) -> None:
+        if (gap := self.minsize - self._pool_size) > 0:
+            await self._init_connections(gap, strict=strict)
 
     @asynccontextmanager
     async def connection(self) -> AsyncIterator[Connection]:
@@ -189,23 +229,16 @@ class Pool:
 
         async with self._sem:
             async with self._lock:
-                if not self._free_connections:
-                    conns, maxsize = self.connections, self.maxsize
-                    avail = maxsize - conns
-                    if (maxsize - conns) < 0:
-                        msg = (
-                            f"the number of pool connections ({conns}) "
-                            f"exceeds the pool maxsize ({maxsize}) for {self}"
-                        )
-                        raise AsynchPoolError(msg)
-                    to_create = min(self.minsize, avail)
-                    await self._init_connections(to_create)
                 conn = await self._acquire_connection()
             try:
                 yield conn
             finally:
                 async with self._lock:
-                    await self._release_connection(conn)
+                    try:
+                        await self._release_connection(conn)
+                    except AsynchPoolError as e:
+                        logger.warning(e)
+                    await self._ensure_minsize_connections(strict=True)
 
     async def startup(self) -> "Pool":
         """Initialise the pool.
@@ -213,12 +246,17 @@ class Pool:
         When entering the context,
         the pool get filled with connections
         up to the pool `minsize` value.
+
+        :return: a pool object with `minsize` opened connections
+        :rtype: Pool
         """
 
         async with self._lock:
             if self._opened:
                 return self
-            await self._init_connections(self.minsize)
+            # If we cannot create the minsize connections here,
+            # the Pool does not meet the minsize requirement.
+            await self._init_connections(self.minsize, strict=True)
             self._opened = True
             if self._closed:
                 self._closed = False
@@ -241,39 +279,3 @@ class Pool:
                 await conn.close()
             self._opened = False
             self._closed = True
-
-
-async def create_pool(
-    minsize: int = constants.POOL_MIN_SIZE,
-    maxsize: int = constants.POOL_MAX_SIZE,
-    loop: Optional[asyncio.AbstractEventLoop] = None,
-    **kwargs,
-) -> Pool:
-    """Returns an initiated connection pool.
-
-    The initiated pool means it is filled with `minsize` connections.
-
-    Equivalent to:
-    1. pool = Pool(...)
-    2. await pool.startup()
-    3. return pool
-
-    Do not forget to `await pool.shutdown()` for resource clean-up.
-
-    :param minsize int: the minimum number of connections in the pool
-    :param maxsize int: the maximum number of connections in the pool
-    :param loop Optional[asyncio.AbstractEventLoop]: an event loop (asyncio.get_running_loop() by default)
-    :param kwargs dict: connection settings
-
-    :return: a connection pool object
-    :rtype: Pool
-    """
-
-    pool = Pool(
-        minsize=minsize,
-        maxsize=maxsize,
-        loop=loop,
-        **kwargs,
-    )
-    await pool.startup()
-    return pool