+++ tests/test_pool.py
@@ -4,9 +4,14 @@ from typing import Any
 import pytest
 
 from asynch.connection import Connection
+from asynch.errors import AsynchPoolError
 from asynch.pool import Pool
 from asynch.proto import constants
 from asynch.proto.models.enums import PoolStatus
+
+
+def _get_pool_size(pool: Pool) -> int:
+    return pool.acquired_connections + pool.free_connections
 
 
 @pytest.mark.asyncio
@@ -56,25 +61,25 @@ async def test_pool_connection_attributes(config):
     pool = Pool(dsn=config.dsn)
     assert pool.minsize == constants.POOL_MIN_SIZE
     assert pool.maxsize == constants.POOL_MAX_SIZE
-    assert pool.connections == 0
+    assert _get_pool_size(pool) == 0
     assert pool.free_connections == 0
     assert pool.acquired_connections == 0
 
     async with pool:
-        assert pool.connections == constants.POOL_MIN_SIZE
+        assert _get_pool_size(pool) == constants.POOL_MIN_SIZE
         assert pool.free_connections == constants.POOL_MIN_SIZE
         assert pool.acquired_connections == 0
 
         async with pool.connection():
-            assert pool.connections == constants.POOL_MIN_SIZE
+            assert _get_pool_size(pool) == constants.POOL_MIN_SIZE
             assert pool.free_connections == 0
             assert pool.acquired_connections == constants.POOL_MIN_SIZE
 
-        assert pool.connections == constants.POOL_MIN_SIZE
+        assert _get_pool_size(pool) == constants.POOL_MIN_SIZE
         assert pool.free_connections == constants.POOL_MIN_SIZE
         assert pool.acquired_connections == 0
 
-    assert pool.connections == 0
+    assert _get_pool_size(pool) == 0
     assert pool.free_connections == 0
     assert pool.acquired_connections == 0
 
@@ -184,3 +189,52 @@ async def test_pool_concurrent_connection_management(get_tcp_connections):
         assert noc == init_tcps
 
     assert selectees == answers
+
+
+@pytest.mark.asyncio
+async def test_pool_broken_connection_handling():
+    async def _get_answer(pool: Pool, *, raise_exc: bool = True):
+        async with pool.connection() as conn_ctx:
+            assert pool.free_connections == 0
+            assert pool.acquired_connections == 1
+
+            async with conn_ctx.cursor() as cur:
+                if raise_exc:
+                    raise AsynchPoolError("good bye")
+                await cur.execute("SELECT 21 + 21;")
+                ret = await cur.fetchone()
+                assert ret == 42
+                return ret
+
+    min_size, max_size = 1, 1
+    pool = Pool(minsize=min_size, maxsize=max_size)
+    async with pool:
+        async with pool.connection() as conn:
+            await conn.ping()
+
+            # he connection is invalidated
+            await conn.close()
+            with pytest.raises(ConnectionError):
+                await conn.ping()
+
+            # but does not influence the pool state
+            assert pool.free_connections == 0
+            assert pool.acquired_connections == 1
+
+        # when leaving the connection context,
+        # the pool should ensure its consistency
+        assert pool.free_connections == 1
+        assert pool.acquired_connections == 0
+
+        async with pool.connection() as conn:
+            await conn.ping()
+            assert pool.free_connections == 0
+            assert pool.acquired_connections == 1
+
+        seq = list(range(10))
+        tasks = [asyncio.create_task(_get_answer(pool=pool, raise_exc=bool(i % 2))) for i in seq]
+        # no blockade and no inconsistency
+        await asyncio.gather(*tasks, return_exceptions=True)
+
+        assert pool.free_connections == 1
+        assert pool.acquired_connections == 0